requirements.txt
flask
spacy
flask-caching
flask-limiter
flasgger
python-dotenv
